Calc Boot-Up
This adds a boot up process of what Calculator you want to use. It is
like when you have a game with multiple games on it and it askes you
what game you want to play. MAIN IS NOW IN CALCRUNNER. Run CalcRuner if
you want to run Chem calc or you could just add a main to chem calc if
you wish. Happy Coding!